requirements.txt
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
